+++ src/lib.rs
@@ -413,7 +413,7 @@ macro_rules! fixed {
             $s_nbits_m1:expr, $s_nbits_m2:expr, $s_nbits_m3:expr, $s_nbits_m4:expr
         ),
         $nbytes:expr, $bytes_val:expr, $rev_bytes_val:expr, $be_bytes:expr, $le_bytes:expr,
-        $UFixed:ident, $UInner:ident, $Signedness:tt,
+        $IFixed:ident, $UFixed:ident, $UInner:ident, $Signedness:tt,
         $LeEqU_C0:tt, $LeEqU_C1:tt, $LeEqU_C2:tt, $LeEqU_C3:tt,
         $HasDouble:tt, $Double:ident, $DoubleInner:ty, $s_nbits_2:expr
     ) => {
@@ -424,7 +424,7 @@ macro_rules! fixed {
                 $s_nbits_m1, $s_nbits_m2, $s_nbits_m3, $s_nbits_m4
             ),
             $nbytes, $bytes_val, $rev_bytes_val, $be_bytes, $le_bytes,
-            $UFixed[stringify!($UFixed)], $UInner, $Signedness,
+            $IFixed[stringify!($IFixed)], $UFixed[stringify!($UFixed)], $UInner, $Signedness,
             $LeEqU_C0, $LeEqU_C1, $LeEqU_C2, $LeEqU_C3,
             $HasDouble, $Double[stringify!($Double)], $DoubleInner, $s_nbits_2
         }
@@ -436,7 +436,7 @@ macro_rules! fixed {
             $s_nbits_m1:expr, $s_nbits_m2:expr, $s_nbits_m3:expr, $s_nbits_m4:expr
         ),
         $nbytes:expr, $bytes_val:expr, $rev_bytes_val:expr, $be_bytes:expr, $le_bytes:expr,
-        $UFixed:ident[$s_ufixed:expr], $UInner:ident, $Signedness:tt,
+        $IFixed:ident[$s_ifixed:expr], $UFixed:ident[$s_ufixed:expr], $UInner:ident, $Signedness:tt,
         $LeEqU_C0:tt, $LeEqU_C1:tt, $LeEqU_C2:tt, $LeEqU_C3:tt,
         $HasDouble:tt, $Double:ident[$s_double:expr], $DoubleInner:ty, $s_nbits_2:expr
     ) => {
@@ -548,7 +548,7 @@ assert_eq!(two_point_75.to_string(), \"2.8\");
         fixed_no_frac! {
             $Fixed[$s_fixed]($Inner[$s_inner], $LeEqU, $UNbits, $s_nbits, $s_nbits_m1, $s_nbits_m2),
             $nbytes, $bytes_val, $rev_bytes_val, $be_bytes, $le_bytes,
-            $UFixed[$s_ufixed], $UInner, $Signedness,
+            $IFixed[$s_ifixed], $UFixed[$s_ufixed], $UInner, $Signedness,
             $HasDouble, $Double[$s_double], $DoubleInner, $s_nbits_2
         }
         // inherent methods that require Frac bounds, and cannot be const
@@ -568,7 +568,7 @@ fixed! {
     "An eight",
     FixedU8(u8, LeEqU8, U8, "8", "7", "6", "5", "4"),
     1, "0x12", "0x12", "[0x12]", "[0x12]",
-    FixedU8, u8, Unsigned,
+    FixedI8, FixedU8, u8, Unsigned,
     U8, U7, U6, U5,
     True, FixedU16, u16, "16"
 }
@@ -576,7 +576,7 @@ fixed! {
     "A 16",
     FixedU16(u16, LeEqU16, U16, "16", "15", "14", "13", "12"),
     2, "0x1234", "0x3412", "[0x12, 0x34]", "[0x34, 0x12]",
-    FixedU16, u16, Unsigned,
+    FixedI16, FixedU16, u16, Unsigned,
     U16, U15, U14, U13,
     True, FixedU32, u32, "32"
 }
@@ -584,7 +584,7 @@ fixed! {
     "A 32",
     FixedU32(u32, LeEqU32, U32, "32", "31", "30", "29", "28"),
     4, "0x1234_5678", "0x7856_3412", "[0x12, 0x34, 0x56, 0x78]", "[0x78, 0x56, 0x34, 0x12]",
-    FixedU32, u32, Unsigned,
+    FixedI32, FixedU32, u32, Unsigned,
     U32, U31, U30, U29,
     True, FixedU64, u64, "64"
 }
@@ -594,7 +594,7 @@ fixed! {
     8, "0x1234_5678_9ABC_DE0F", "0x0FDE_BC9A_7856_3412",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0x0F]",
     "[0x0F, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
-    FixedU64, u64, Unsigned,
+    FixedI64, FixedU64, u64, Unsigned,
     U64, U63, U62, U61,
     True, FixedU128, u128, "128"
 }
@@ -607,7 +607,7 @@ fixed! {
      0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08]",
     "[0x08, 0x07, 0x06, 0x05, 0x04, 0x03, 0x02, 0x01, \
      0xF0, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
-    FixedU128, u128, Unsigned,
+    FixedI128, FixedU128, u128, Unsigned,
     U128, U127, U126, U125,
     False, FixedU128, u128, "128"
 }
@@ -615,7 +615,7 @@ fixed! {
     "An eight",
     FixedI8(i8, LeEqU8, U8, "8", "7", "6", "5", "4"),
     1, "0x12", "0x12", "[0x12]", "[0x12]",
-    FixedU8, u8, Signed,
+    FixedI8, FixedU8, u8, Signed,
     U7, U6, U5, U4,
     True, FixedI16, i16, "16"
 }
@@ -623,7 +623,7 @@ fixed! {
     "A 16",
     FixedI16(i16, LeEqU16, U16,"16", "15", "14", "13", "12"),
     2, "0x1234", "0x3412", "[0x12, 0x34]", "[0x34, 0x12]",
-    FixedU16, u16, Signed,
+    FixedI16, FixedU16, u16, Signed,
     U15, U14, U13, U12,
     True, FixedI32, i32, "32"
 }
@@ -631,7 +631,7 @@ fixed! {
     "A 32",
     FixedI32(i32, LeEqU32, U32, "32", "31", "30", "29", "28"),
     4, "0x1234_5678", "0x7856_3412", "[0x12, 0x34, 0x56, 0x78]", "[0x78, 0x56, 0x34, 0x12]",
-    FixedU32, u32, Signed,
+    FixedI32, FixedU32, u32, Signed,
     U31, U30, U29, U28,
     True, FixedI64, i64, "64"
 }
@@ -641,7 +641,7 @@ fixed! {
     8, "0x1234_5678_9ABC_DE0F", "0x0FDE_BC9A_7856_3412",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0x0F]",
     "[0x0F, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
-    FixedU64, u64, Signed,
+    FixedI64, FixedU64, u64, Signed,
     U63, U62, U61, U60,
     True, FixedI128, i128, "128"
 }
@@ -654,7 +654,7 @@ fixed! {
      0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08]",
     "[0x08, 0x07, 0x06, 0x05, 0x04, 0x03, 0x02, 0x01, \
      0xF0, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
-    FixedU128, u128, Signed,
+    FixedI128, FixedU128, u128, Signed,
     U127, U126, U125, U124,
     False, FixedI128, i128, "128"
 }